requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.8.2
google-generativeai==0.8.3
supabase==2.7.4
python-dotenv==1.0.1
httpx==0.27.2
